Move scripts/repo_gen to cfme/utils/
update imports from cfme/scripting/appliance and
cfme/tests/test_db_migrate and cfme/tests/cli/test_appliance_update

No updates to modules outside of move/import changes
